create order

+++ Models/Order.cs
@@ -8,7 +8,7 @@ namespace debt_collector_api.Models
         public int ExpenseId { get; set; }
         public Expense Expense { get; set; }
         public string Name { get; set; }
-        public int TotalCost { get; set; }
+        public decimal TotalCost { get; set; }
         public List<Debtor> Debtors { get; set; } = [];
         public List<Payer> Payers { get; set; } = [];
     }
